// packages/servequery-cloud/src/types.ts
export type EnvironmentVariables = {
  TOKEN_PATH: string;
  SERVEQUERY_ENV_SECRET: string;
  SERVEQUERY_SERVER_URL: string;
  SERVEQUERY_SUBSCRIPTION_URL: string;
  SERVEQUERY_AUTH_TOKEN: string;
  NODE_TLS_REJECT_UNAUTHORIZED: string;
};